Use Native Texture Formats - color_formats.h:   - `union RGBA16` has had a bug fix. Previously, it erroneously had a 4-byte size, but now it has the proper 2-byte size. - gfx_citro3d_helpers.c and .h:   - Texture padding function names have been improved.   - Texture padding functions have been cleaned up and optimized. They also no longer accept `color_formats.h` types.   - Logging has been improved.   - Texture dimension adjustment is now located here. Previously, it was in `gfx_citro3d.c.` It accepts and returns the new `struct TextureSize` type. - gfx_citro3d.c:   - Now uses native N3DS texture formats when possible. This has the following benefits:     - The TMEM footprint is reduced.     - Loading times are substantially improved, as the most common formats are supported natively and therefore do not require conversion, and sending to TMEM takes less time as they are smaller.   - Now handles texture conversion, instead of `gfx_pc.c`   - Upon uploading a texture, now checks to ensure that the C3D texture was created properly. - gfx_pc.c:   - No longer handles texture conversion. Instead, this is handled by the rendering API, to allow for native format support.   - Added a new macro and set of definitions for defining supported N64 texture types.   - `import_texture` has been cleaned up, now using the definitions mentioned above and generally having much, much cleaner code. - gfx_rendering_api.h:   - Adjusted to include support for native texture formats. - texture_conversion.c and .h:   - Added and used a set of accurate color upscaling macros for 3->8-bit and 3->4-bit conversions. These macros properly maintain their target format's entire value ranges.   - Changed the order of parameters in conversion functions that deal with palettes. Now, the palette is passed adjacent to the other buffers.   - Conversion functions now accept width/height rather than buffer size.   - Added more conversion functions.   - Removed incorrect calls to the BIG_ENDIAN_LOAD macro, as the data is already in the correct format.

+++ src/pc/gfx/color_formats.h
@@ -15,13 +15,13 @@ union RGBA32 {
     uint32_t u32;
 };
 
-// One 16-bit RGBA color
+// One 16-bit RGBA color. Uses u16 bitfields because u8s bloat the struct size to 4U.
 union RGBA16 {
     struct {
-        uint8_t r : 5;
-        uint8_t g : 5;
-        uint8_t b : 5;
-        uint8_t a : 1;
+        uint16_t r : 5;
+        uint16_t g : 5;
+        uint16_t b : 5;
+        uint16_t a : 1;
     };
     uint16_t u16;
 };
